#0: Break HS form into its own component
