+++ website/.vuepress/theme/layouts/Join.vue
@@ -30,13 +30,7 @@
               <h3>I'd love to join <em>The Alliance</em> as a(n)...</h3>
 
               <p>
-                <script>
-                  // Load our hubspot form
-                  hbspt.forms.create({
-                    portalId: '6478338',
-                    formId: '9017bea2-dd0c-402a-8c97-d1b1d7429a00',
-                  });
-                </script>
+                <HSForm form="9017bea2-dd0c-402a-8c97-d1b1d7429a00"></HSForm>
               </p>
             </div>
           </div>
@@ -58,80 +52,13 @@ import Sidebar from '@theme/components/Sidebar.vue';
 import {resolveSidebarItems} from '@parent-theme/util/index.js';
 
 // Lando components
+import Footer from '@theme/components/Footer';
+import HSForm from '@theme/components/HSForm';
 import MadeByTandem from '@theme/components/MadeByTandem';
 import Newsletter from '@theme/components/Newsletter';
-import Footer from '@theme/components/Footer';
-
-// CSS
-const hscss = `
-  .inputs-list label {
-    line-height: 26px;
-  }
-  form {
-    margin: auto;
-    padding: 5em;
-    border: 1px #ed3f7a solid;
-  }
-  .inputs-list:first-child {
-    padding-top: 2em;
-  }
-  .inputs-list .hs-form-checkbox  {
-    padding-bottom: 2em;
-  }
-  .inputs-list .hs-form-checkbox small  {
-    color: #4e6e8e;
-  }
-  .multi-container {
-    padding-bottom: 1em;
-  }
-  .hs-custom-style fieldset input:not([type="image"]):not([type="submit"]):not([type="button"]):not([type="radio"]):not([type="checkbox"]):not([type="file"]), .hs-custom-style>div input:not([type="image"]):not([type="submit"]):not([type="button"]):not([type="radio"]):not([type="checkbox"]):not([type="file"]) {
-    -webkit-writing-mode: horizontal-tb !important;
-    text-rendering: auto;
-    color: initial;
-    letter-spacing: normal;
-    word-spacing: normal;
-    text-transform: none;
-    text-indent: 0px;
-    text-shadow: none;
-    display: inline-block;
-    text-align: start;
-    -webkit-appearance: textfield;
-    background-color: white;
-    -webkit-rtl-ordering: logical;
-    cursor: text;
-    margin: 0em;
-    font: 400 11px system-ui;
-    padding: 1px;
-    border-width: 2px;
-    border-style: inset;
-    border-color: initial;
-    border-image: initial;
-    width: 100%;
-    box-sizing: border-box;
-    padding: 10px 80px 10px 20px;
-    height: 50px;
-    border-radius: 50px;
-    border: 1px solid #ccc;
-    font-size: 16px;
-  }
-  .hs-available_for_contract_work label.hs-form-booleancheckbox-display {
-    margin-top: -38px;
-    font-size: 1em;
-  }
-  .hs-available_for_contract_work label.hs-form-booleancheckbox-display span {
-    color: #4e6e8e;
-  }
-  @media (max-width: 750px) {
-    form {
-      margin: auto;
-      padding: 0;
-      border: 0;
-    }
-  }
-`;
 
 export default {
-  components: {Home, Sidebar, Navbar, MadeByTandem, Newsletter, Footer},
+  components: {Home, HSForm, Sidebar, Navbar, MadeByTandem, Newsletter, Footer},
   data() {
     return {
       isSidebarOpen: false,
@@ -187,17 +114,6 @@ export default {
     this.$router.afterEach(() => {
       this.isSidebarOpen = false;
     });
-
-    // Let's inject some CSS to make things look a bit nicer but add some
-    // nasty recursion just to make sure everthing is in order
-    const iid = setInterval(() => {
-      if (document.getElementById('hs-form-iframe-0') !== null) {
-        const joinHTML = document.getElementById('hs-form-iframe-0');
-        joinHTML.contentDocument.body.innerHTML += `<style>${hscss}</style>`;
-        joinHTML.style.height = '1000px';
-        window.clearInterval(iid);
-      }
-    }, 250);
   },
   methods: {
     toggleSidebar(to) {
